pyproject.toml
[project]
name = "vaastubbox"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "langchain>=1.2.3",
    "langchain-openai>=1.1.7",
    "pandas>=2.3.3",
    "pdfplumber>=0.11.9",
    "pillow>=12.1.0",
    "streamlit>=1.52.2",
]
